lesson5/requirements.txt
fastapi==0.89.1
uvicorn==0.22.0
python-dotenv==1.0.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
opentelemetry-instrumentation-fastapi==0.40b0
opentelemetry-instrumentation-requests==0.40b0
opentelemetry-exporter-jaeger==1.19.0
requests==2.31.0
